backend_rentelektro/pyproject.toml
[project]
name = "backend-rentelektro"
version = "0.1.0"
description = "Backend service for the RentElektro project."
authors = [
  { name = "RentElektro Team" },
]
requires-python = ">=3.12,<4.0"
dependencies = [
  "fastapi>=0.135,<0.136",
  "uvicorn>=0.41,<0.42",
  "sqlalchemy>=2.0.46,<2.1",
  "alembic>=1.16,<2",
  "gunicorn>=23,<24",
  "pyjwt>=2.12.1,<3",
  "bcrypt==4.0.1",
  "passlib==1.7.4",
  "prometheus-fastapi-instrumentator>=7.1,<8",
  "asyncpg>=0.30,<1",
  "environ>=1.0,<2",
  "pydantic[email]>=2.11,<3",
  "pydantic-settings>=2.12,<3",
  "psycopg2-binary>=2.9.11,<3",
  "python-multipart>=0.0.20,<0.1",
  "httpx>=0.28,<1",
  "nanoid (>=2.0.0,<3.0.0)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.3,<10"
ruff = ">=0.12,<0.13"
pre-commit = ">=4,<5"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [
  "migrations",
]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
